moved knexfile
