+++ backend/.env
@@ -2,13 +2,6 @@ SECRET_KEY = eyJhbGciOiJIUzI1NiJ9.eyJSb2xlIjoiQWRtaW4iLCJJc3N1ZXIiOiJJc3N1ZXIiLC
 PORT=4000
 POSTGRES_HOST=postgres
 POSTGRES_USER=postgres
-POSTGRES_DB=task-tracker
-POSTGRESS_PASSWORD="0896"
+POSTGRES_DB=postgres
+POSTGRES_PASSWORD='1234'
 POSTGRESS_PORT=5432
-
-
-
-
-# DB_NAME=task-tracker
-# DB_USER=postgres
-# DB_PASSWORD="0896"
