+++ backend/index.js
@@ -11,9 +11,9 @@ app.use(cors());
 app.use(express.json());
 app.use("/api", router);
 
-app.get('/checkServer',(req,res)=>{
-    res.json({message:'docker server launched'})
-})
+app.get("/api/checkServer", (req, res) => {
+  res.json({ message: "!!!!docker server launched!!!" });
+});
 
 const start = async () => {
   try {
