+++ backend/knexfile.js
@@ -1,17 +1,21 @@
 import {} from "dotenv/config";
 
+// console.log(process.env);
 export const development = {
   client: "postgresql",
   connection: {
     database: process.env.POSTGRES_DB,
     user: process.env.POSTGRES_USER,
-    password: process.env.POSTGRESS_PASSWORD,
+    password: process.env.POSTGRES_PASSWORD,
   },
   pool: {
     min: 2,
     max: 10,
   },
   migrations: {
-    tableName: "knex_migrations",
+    directory: "./db/migrations",
+  },
+  seeds: {
+    directory: "./db/seeds",
   },
 };
