+++ backend/package.json
@@ -1,23 +1,23 @@
 {
-    "scripts": {
-        "serve": "cross-env NODE_ENV=development nodemon ./index.js",
-        "knex:make": "cross-env NODE_ENV=development knex migrate:make --knexfile db/knexfile.js",
-        "knex:migrate": "cross-env NODE_ENV=development knex migrate:latest --knexfile db/knexfile.js",
-        "knex:rollback": "cross-env NODE_ENV=development knex migrate:rollback --knexfile db/knexfile.js",
-        "knex:seed": "cross-env NODE_ENV=development knex seed:run --knexfile db/knexfile.js"
-    },
-    "dependencies": {
-        "bcrypt": "^5.1.1",
-        "cors": "^2.8.5",
-        "dotenv": "^16.3.1",
-        "express": "^4.18.2",
-        "jsonwebtoken": "^9.0.2",
-        "knex": "^3.1.0",
-        "pg": "^8.11.3"
-    },
-    "devDependencies": {
-        "cross-env": "^7.0.3",
-        "nodemon": "^3.0.2"
-    },
-    "type": "module"
+  "scripts": {
+    "start": "cross-env NODE_ENV=development nodemon ./index.js",
+    "knex:make": "cross-env NODE_ENV=development knex migrate:make --knexfile ./knexfile.js",
+    "knex:migrate": "cross-env NODE_ENV=development knex migrate:latest --knexfile ./knexfile.js",
+    "knex:rollback": "cross-env NODE_ENV=development knex migrate:rollback --knexfile ./knexfile.js",
+    "knex:seed": "cross-env NODE_ENV=development knex seed:run --knexfile ./knexfile.js"
+  },
+  "dependencies": {
+    "bcrypt": "^5.1.1",
+    "cors": "^2.8.5",
+    "dotenv": "^16.3.1",
+    "express": "^4.18.2",
+    "jsonwebtoken": "^9.0.2",
+    "knex": "^3.1.0",
+    "pg": "^8.11.3"
+  },
+  "devDependencies": {
+    "cross-env": "^7.0.3",
+    "nodemon": "^3.0.2"
+  },
+  "type": "module"
 }
